mark auth pages as client components
