+++ app/login/page.jsx
@@ -1,36 +1,32 @@
+'use client'
 import { useState } from 'react';
-import { useRouter } from 'next/router';
+import { signIn } from 'next-auth/react';
+import { useRouter } from 'next/navigation';
 
-export default function Register() {
+export default function Login() {
   const router = useRouter();
   const [email, setEmail] = useState('');
   const [password, setPassword] = useState('');
-  const [confirm, setConfirm] = useState('');
   const [error, setError] = useState('');
 
   const handleSubmit = async e => {
     e.preventDefault();
-    const res = await fetch('/api/auth/register', {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({ email, password, confirm })
+    const res = await signIn('credentials', {
+      redirect: false,
+      email,
+      password
     });
-    if (res.ok) {
-      router.push('/login');
-    } else {
-      const data = await res.json();
-      setError(data.message || 'Errore');
-    }
+    if (res.ok) router.push('/');
+    else setError('Credenziali non valide');
   };
 
   return (
     <div className="auth-container">
-      <h1>Registrati</h1>
+      <h1>Login</h1>
       <form onSubmit={handleSubmit}>
         <input type="email" value={email} onChange={e=>setEmail(e.target.value)} placeholder="Email" required />
         <input type="password" value={password} onChange={e=>setPassword(e.target.value)} placeholder="Password" required />
-        <input type="password" value={confirm} onChange={e=>setConfirm(e.target.value)} placeholder="Conferma Password" required />
-        <button type="submit">Registrati</button>
+        <button type="submit">Login</button>
       </form>
       {error && <p style={{color:'red'}}>{error}</p>}
     </div>
