+++ app/register/page.jsx
@@ -1,31 +1,37 @@
+'use client'
 import { useState } from 'react';
-import { signIn } from 'next-auth/react';
-import { useRouter } from 'next/router';
+import { useRouter } from 'next/navigation';
 
-export default function Login() {
+export default function Register() {
   const router = useRouter();
   const [email, setEmail] = useState('');
   const [password, setPassword] = useState('');
+  const [confirm, setConfirm] = useState('');
   const [error, setError] = useState('');
 
   const handleSubmit = async e => {
     e.preventDefault();
-    const res = await signIn('credentials', {
-      redirect: false,
-      email,
-      password
+    const res = await fetch('/api/auth/register', {
+      method: 'POST',
+      headers: { 'Content-Type': 'application/json' },
+      body: JSON.stringify({ email, password, confirm })
     });
-    if (res.ok) router.push('/');
-    else setError('Credenziali non valide');
+    if (res.ok) {
+      router.push('/login');
+    } else {
+      const data = await res.json();
+      setError(data.message || 'Errore');
+    }
   };
 
   return (
     <div className="auth-container">
-      <h1>Login</h1>
+      <h1>Registrati</h1>
       <form onSubmit={handleSubmit}>
         <input type="email" value={email} onChange={e=>setEmail(e.target.value)} placeholder="Email" required />
         <input type="password" value={password} onChange={e=>setPassword(e.target.value)} placeholder="Password" required />
-        <button type="submit">Login</button>
+        <input type="password" value={confirm} onChange={e=>setConfirm(e.target.value)} placeholder="Conferma Password" required />
+        <button type="submit">Registrati</button>
       </form>
       {error && <p style={{color:'red'}}>{error}</p>}
     </div>
