--- pages/_document.js
@@ -1,23 +1,0 @@
-import Document, { Html, Head, Main, NextScript } from 'next/document';
-
-class MyDocument extends Document {
-  static async getInitialProps(ctx) {
-    const initialProps = await Document.getInitialProps(ctx);
-    return { ...initialProps, locale: ctx.locale || 'it' };
-  }
-
-  render() {
-    const { locale } = this.props;
-    return (
-      <Html lang={locale}>
-        <Head />
-        <body data-lang={locale}>
-          <Main />
-          <NextScript />
-        </body>
-      </Html>
-    );
-  }
-}
-
-export default MyDocument;
